Timer values are collected in nano seconds. This would phase off the projected percentile calculations from codehaul metrics

+++ loader-core/src/main/java/com/open/perf/core/SyncFunctionExecutor.java
@@ -78,7 +78,7 @@ public class SyncFunctionExecutor {
     }
 
     //Nano Seconds
-    public double getExecutionTime() {
+    public long getExecutionTime() {
         return (this.endTime - this.startTime);
     }
